fix: resolve breaking tests
