+++ src/containers/Tables/BidTable/BidTable.types.tsx
@@ -1,6 +1,6 @@
 //- Library Imports
 import { ethers } from 'ethers';
-import { Account, Domain } from 'lib/types';
+import { Domain } from 'lib/types';
 import { ConvertedTokenInfo } from '@zero-tech/zns-sdk';
 
 //- Constants Imports
@@ -26,25 +26,13 @@ export const ActionKeys = {
 	CANCEL_BID: OptionTitle.CANCEL_BID,
 };
 
-export const getTableActions = (
-	account: string | null | undefined,
-	owner: Account,
-) => {
-	const bidTableActions = [
-		{
-			icon: <DollarSign />,
-			title: ActionKeys.REBID,
-		},
-		{
-			icon: <X />,
-			title: ActionKeys.CANCEL_BID,
-		},
-	];
-
-	const filteredActions =
-		String(owner) === account
-			? bidTableActions.filter((item) => item.title !== ActionKeys.REBID)
-			: bidTableActions;
-
-	return filteredActions;
-};
+export const bidTableActions = [
+	{
+		icon: <DollarSign />,
+		title: ActionKeys.REBID,
+	},
+	{
+		icon: <X />,
+		title: ActionKeys.CANCEL_BID,
+	},
+];
