+++ src/containers/Tables/BidTable/BidTable.utils.ts
@@ -1,3 +1,7 @@
+//- Library Imports
+import { Account } from 'lib/types';
+import { ActionKeys, bidTableActions } from './BidTable.types';
+
 export const handleDomainNameWidth = (dimensions: number) => {
 	const isMobileM = dimensions <= 400;
 	const isMobileL = dimensions <= 481;
@@ -12,4 +16,16 @@ export const handleDomainNameWidth = (dimensions: number) => {
 	} else {
 		return 420;
 	}
+};
+
+export const getTableActions = (
+	account: string | null | undefined,
+	owner: Account,
+) => {
+	const filteredActions =
+		String(owner) === account
+			? bidTableActions.filter((item) => item.title !== ActionKeys.REBID)
+			: bidTableActions;
+
+	return filteredActions;
 };
